+++ sync-for-expenses/CodatSyncExpenses/CodatSyncExpensesSDK.cs
@@ -121,10 +121,10 @@ namespace CodatSyncExpenses
         };
 
         private const string _language = "csharp";
-        private const string _sdkVersion = "3.3.0";
-        private const string _sdkGenVersion = "2.169.0";
+        private const string _sdkVersion = "3.3.1";
+        private const string _sdkGenVersion = "2.173.0";
         private const string _openapiDocVersion = "prealpha";
-        private const string _userAgent = "speakeasy-sdk/csharp 3.3.0 2.169.0 prealpha Codat.Sync.Expenses";
+        private const string _userAgent = "speakeasy-sdk/csharp 3.3.1 2.173.0 prealpha Codat.Sync.Expenses";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private ISpeakeasyHttpClient _securityClient;