+++ sync-for-expenses/RELEASES.md
@@ -121,3 +121,13 @@ Based on:
 - [csharp v3.3.0] sync-for-expenses
 ### Releases
 - [NuGet v3.3.0] https://www.nuget.org/packages/Codat.Sync.Expenses/3.3.0 - sync-for-expenses
+
+## 2023-10-30 01:27:22
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v3.3.1] sync-for-expenses
+### Releases
+- [NuGet v3.3.1] https://www.nuget.org/packages/Codat.Sync.Expenses/3.3.1 - sync-for-expenses